+++ hutool-core/src/test/java/cn/hutool/core/math/CalculatorTest.java
@@ -1,7 +1,6 @@
 package cn.hutool.core.math;
 
 import org.junit.Assert;
-import org.junit.Ignore;
 import org.junit.Test;
 
 public class CalculatorTest {
@@ -41,5 +40,12 @@ public class CalculatorTest {
 	public void conversationTest6() {
 		final double conversion = Calculator.conversion("-((2.12-2) * 100)");
 		Assert.assertEquals(-1D * (2.12 - 2) * 100, conversion, 2);
+	}
+
+	@Test
+	public void conversationTest7() {
+		//https://gitee.com/dromara/hutool/issues/I4KONB
+		final double conversion = Calculator.conversion("((-2395+0) * 0.3+140.24+35+90)/30");
+		Assert.assertEquals(-15.11, conversion, 2);
 	}
 }