+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * 【core   】     修复NamingCase中转换下划线字母+数字转换问题（issue#2070@Github）
 * 【core   】     修复split空判断不一致问题（pr#496@Gitee）
 * 【crypto 】     修复SM2.getDHex()前导0丢失,然后导致获取密钥错误（pr#2073@Github）
+* 【core   】     修复关于Calculator.conversion()方法EmptyStackException的bug（pr#2076@Github）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.7.18 (2021-12-25)